Options, Meta APIs: Correct some inline docs for `update_post_meta` and `delete_post_meta`.
In `update_post_meta` and `delete_post_meta`, code was added to fetch the parent post if these functions were called for a revision post ID.  In [9252], the code was apparently copied from `add_post_meta`, and the inline comment describing the new block of code was not updated to mention either "updated" or "deleted" respectively.  This change corrects the comment in both functions to reflect the current action instead of "added".

Props thelovekesh, rehanali.
Fixes #54835.

git-svn-id: https://develop.svn.wordpress.org/trunk@52660 602fd350-edb4-49c9-b593-d223f7449a82