Database: Use generic references to a database server in `ms_not_installed()`.
This replaces the references to “MySQL” with “your host’s database server” on the “Error establishing a database connection” page when Multisite could not be loaded. The generic “database server” term can refer to MySQL or MariaDB. Additionally, this brings some consistency with a similar error message in `wpdb::db_connect()`.

Follow-up to [52367], [52423].

Props tj692, hansjovisyoast, tobifjellner.
Fixes #55701.

git-svn-id: https://develop.svn.wordpress.org/trunk@53394 602fd350-edb4-49c9-b593-d223f7449a82